Add data

See merge request jorge.paravicini/jee_webshop!20

+++ src/main/kotlin/com/jorgeparavicini/springwebshop/services/ProductService.kt
@@ -14,7 +14,7 @@ interface ProductService : Service<Product, ProductDTO, ProductRepository> {
         minPrice: Float?,
         maxPrice: Float?,
         maxShippingPrice: Float?,
-        minRating: Int?
+        minRating: Float?
     ): ProductListFilterInfo
 
     fun getAllRelatedProducts(productId: Long): Iterable<RelatedProductDTO>
